requirements.txt
tldextract
newsapi-python
nltk
bs4
flask
gunicorn